Remove main.rs
